requirements.txt
python-dotenv==0.10.3
python-telegram-bot==12.0.0
requests==2.22.0
